Bug fixes whilst user testing

+++ ReviewService/Controllers/CustomerReviewController.cs
@@ -13,13 +13,14 @@ using ReviewService.Models;
 namespace ReviewService.Controllers
 {
     [ApiController]
-    [Route("api/[controller]")]
+    [Route("api/Customer")]
     [Authorize(Policy = "CustomerOnly")]
     public class CustomerReviewController : ControllerBase
     {
         private readonly ILogger<CustomerReviewController> _logger;
         private readonly IReviewRepository _reviewRepo;
         private readonly IMapper _mapper;
+        private string authId, clientId;
 
         public CustomerReviewController(ILogger<CustomerReviewController> logger, IReviewRepository reviewRepo, IMapper mapper)
         {
@@ -28,16 +29,31 @@ namespace ReviewService.Controllers
             _mapper = mapper;
         }
 
+        private void GetTokenDetails()
+        {
+            authId = User
+                .Claims
+                .FirstOrDefault(c => c.Type == "sub")?.Value;
+            clientId = User
+                .Claims
+                .FirstOrDefault(c => c.Type == "client_id")?.Value;
+        }
+
         [HttpGet]
         public async Task<IActionResult> Get(int? customerId, int? productId)
         {
+            if (productId != null && productId > 0 && (customerId == null || customerId < 1))
+            {
+                return Ok(_mapper.Map<List<ReviewDto>>(await _reviewRepo.GetReviewsByProductId(productId ?? 0, visible: true)));
+            }
+            GetTokenDetails();
+            if (! await _reviewRepo.ValidAuthId(customerId??0, authId))
+            {
+                return Forbid();
+            }
             if (customerId != null && customerId > 0 && (productId == null || productId < 1))
             {
                 return Ok(_mapper.Map<List<ReviewDto>>(await _reviewRepo.GetReviewsByCustomerId(customerId??0, visible: true)));
-            }
-            if (productId != null && productId > 0 && (customerId == null || customerId < 1))
-            {
-                return Ok(_mapper.Map<List<ReviewDto>>(await _reviewRepo.GetReviewsByProductId(productId??0, visible: true)));
             }
             var review = _mapper.Map<ReviewDto>(_reviewRepo.GetReview(customerId??0, productId??0, staff: false));
             if (review != null)
@@ -61,6 +77,11 @@ namespace ReviewService.Controllers
 
         private async Task<IActionResult> CreateOrEdit([FromBody] ReviewDto reviewDto)
         {
+            GetTokenDetails();
+            if (!await _reviewRepo.ValidAuthId(reviewDto.CustomerId, authId))
+            {
+                return Forbid();
+            }
             if (!ValidReview(reviewDto))
             {
                 return UnprocessableEntity();
@@ -99,6 +120,10 @@ namespace ReviewService.Controllers
         {
             //if date is over 7 days old, or a future date, set date to now (7 days chosen arbitrarily as it would 
             //likely be a business decision above my position)
+            var now = DateTime.Now.Ticks;
+            var orderTicks = orderDate.Ticks;
+            var difference = now - orderTicks;
+            var limit = TimeSpan.TicksPerDay * 7;
             if (DateTime.Now.Ticks - orderDate.Ticks > (TimeSpan.TicksPerDay * 7) || orderDate > DateTime.Now)
             {
                 return DateTime.Now;
@@ -109,6 +134,11 @@ namespace ReviewService.Controllers
         [HttpDelete]
         public async Task<IActionResult> Delete(int customerId, int productId)
         {
+            GetTokenDetails();
+            if (!await _reviewRepo.ValidAuthId(customerId, authId))
+            {
+                return Forbid();
+            }
             if (await _reviewRepo.DeleteReview(customerId, productId))
             {
                 return Ok();
